Time: 1 ms (99.46%), Space: 41.9 MB (64.90%) - LeetHub

+++ 3110-score-of-a-string/3110-score-of-a-string.java
@@ -3,7 +3,7 @@ class Solution {
         char ch[] = s.toCharArray();
         int sum =0;
         for(int i=1; i<ch.length;i++){
-            sum+=Math.abs((int)ch[i]-(int)ch[i-1]);
+            sum+=Math.abs(ch[i]-ch[i-1]);
         }
         return sum;
     }
